colocando codigo fonte de java para DevOps

+++ src/main/java/br/com/project/prontpet/dtos/AppointmentRequest.java
@@ -7,7 +7,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 
@@ -34,7 +34,7 @@ public record AppointmentRequest (
         LocalDateTime appointmentDate,
 
         @NotNull(message =  "updatedWeight is required")
-        Double updatedWeight
+        BigDecimal updatedWeight
 ){
     public Appointment toEntity(){
         return Appointment.builder()
